Fix cleaning of peer dependencies (#27)
* Fix cleaning of peer dependencies

Peer dependencies with sub folders where not cleaned so far. The
reason for that is that the generated glob was only about the direct
sub folder and deeper structures were ignored.

* Significant improve cleanup performance

Currently each target created two glob patterns. Those add up to quite
a lot of patterns that must all be checked against a lot of files.
This replaces those with a single glob pattern that combines all
targets. That reduces the CPU time significantly.